Added test for correct time
